// apps/api/src/gym-sessions/dto/start-gym-session.dto.d.ts
export declare class StartGymSessionDto {
    workoutPlanId?: string;
    workoutDayIndex?: number;
    adHocExercises?: {
        exerciseId: string;
        targetSets: number;
        targetReps: number;
        targetWeight?: number;
        restSeconds?: number;
    }[];
}
//# sourceMappingURL=start-gym-session.dto.d.ts.map